requirements.txt
Django==1.6.2
Pillow==2.4.0
djangorestframework==2.3.13
gunicorn==18.0
